feat: split cardlist file on two(pagination for all pokemons and card list)
